Initial work integrating cross-server client code
Darmok uses a RemoteUser and LocalUser classes to manage different users. Server code not here
yet, and still code work to do. I'm sure there are bugs, this isn't tested.

The overall goal is to allow darmok to act as either a server or a client. Clients connect
and pass messages/commands to the server and the server updates all clients. The server
will primarily be responsible for ensuring each client user has proper access, i.e. channel
bans etc will be managed on the server.